feat(matching): updates to commit
refs: #1037
