+++ package.json
@@ -72,8 +72,5 @@
     "lint-staged": "^13.1.0",
     "semantic-release": "^20.0.2",
     "typescript": "^4.9.5"
-  },
-  "dependencies": {
-    "jaro-winkler": "^0.2.8"
   }
 }
