+++ packages/api/package.json
@@ -75,6 +75,7 @@
     "@types/cors": "^2.8.12",
     "@types/express": "^4.17.14",
     "@types/express-http-proxy": "^1.6.3",
+    "@types/jaro-winkler": "^0.2.2",
     "@types/jest": "^29.5.1",
     "@types/json-stringify-safe": "^5.0.0",
     "@types/jsonwebtoken": "^8.5.9",
