+++ packages/api/src/command/medical/patient/get-patient.ts
@@ -3,12 +3,14 @@ import { Op, Transaction } from "sequelize";
 import NotFoundError from "../../../errors/not-found";
 import { FacilityModel } from "../../../models/medical/facility";
 import { OrganizationModel } from "../../../models/medical/organization";
-import { Patient, PatientData, PatientModel, splitName } from "../../../models/medical/patient";
+import { Patient, PatientData, PatientModel } from "../../../models/medical/patient";
 import { capture } from "../../../shared/notifications";
 import { Util } from "../../../shared/util";
 import { getFacilities } from "../facility/get-facility";
 import { getOrganizationOrFail } from "../organization/get-organization";
-import jaroWinkler from "jaro-winkler";
+import { calculatePatientSimilarity } from "./calculate-patient-similarity";
+
+const JW_THRESHOLD = 0.95;
 
 export const getPatients = async ({
   facilityId,
@@ -57,8 +59,6 @@ export const getPatientByDemo = async ({
     },
   });
 
-  // TODO: #656 Check for personal identifiers & demo in memory, we were having a hard time to get the query to work with Sequelize
-  // Consider checking this out if trying to move this to the query: https://github.com/sequelize/sequelize/issues/5173
   const matchingPatients = patients.filter(patient => {
     // First, check for an ID match - if it's a match, don't bother checking for demo
     if (
@@ -72,21 +72,9 @@ export const getPatientByDemo = async ({
 
     // Basic Implementation with semi-arbitrary threshold
     const jw_score = calculatePatientSimilarity(patient.data, demo);
-    if (jw_score >= 0.95) {
+    if (jw_score >= JW_THRESHOLD) {
       return patient;
     }
-
-    // If the IDs don't match, or none were provided, check the demo for a match
-    // let demoMatch =
-    //   intersectionWith(splitName(patient.data.firstName), splitName(demo.firstName), isEqual)
-    //     .length > 0 &&
-    //   intersectionWith(splitName(patient.data.lastName), splitName(demo.lastName), isEqual).length >
-    //     0 &&
-    //   intersectionWith(patient.data.address, demo.address, isEqual).length > 0;
-    // if (demoMatch && demo.contact && demo.contact.length > 0) {
-    //   demoMatch = intersectionWith(patient.data.contact, demo.contact, isEqual).length > 0;
-    // }
-    // return demoMatch;
   });
   if (matchingPatients.length === 0) return null;
   if (matchingPatients.length === 1) return matchingPatients[0];
@@ -158,76 +146,4 @@ export const getPatientWithDependencies = async ({
   const facilities = await getFacilities({ cxId, ids: patient.facilityIds });
   const organization = await getOrganizationOrFail({ cxId });
   return { patient, facilities, organization };
-};
-
-/**
- * This function calculates the similarity between two patients using the Jaro-Winkler algorithm.
- * It returns a score between 0 and 1, where 1 means the patients are identical. We calculate scores
- * for the following fields: First Name, Last Name, Address Line 1, Address Line 2, City, State, Country, Zipcode.
- * This function won't be called if gender and DOB are not identical, so that is a givem.
- * @param patient1
- * @param patient2
- * @returns The average of the similarity scores for each field.
- */
-export const calculatePatientSimilarity = (
-  patient1: PatientData,
-  patient2: PatientData
-): number => {
-  let score = 0;
-  let fieldCount = 0;
-  const similarityScores: { [key: string]: [number, string?, string?] } = {};
-
-  const addScore = (field: string, value1: any, value2: any) => {
-    const similarity = jaroWinkler(value1, value2);
-    similarityScores[field] = [similarity, value1, value2];
-    score += similarity;
-    fieldCount += 1;
-  };
-
-  addScore("First Name", patient1.firstName, patient2.firstName);
-  addScore("Last Name", patient1.lastName, patient2.lastName);
-
-  // Calculate similarity for addresses
-  const address1 = patient1.address && patient1.address.length > 0 ? patient1.address[0] : null;
-  const address2 = patient2.address && patient2.address.length > 0 ? patient2.address[0] : null;
-
-  if (address1 && address2) {
-    if (address1.addressLine1 && address2.addressLine1) {
-      addScore("Address Line 1", address1.addressLine1, address2.addressLine1);
-    }
-
-    if (address1.addressLine2 && address2.addressLine2) {
-      addScore("Address Line 2", address1.addressLine2, address2.addressLine2);
-    }
-
-    if (address1.city && address2.city) {
-      addScore("City", address1.city, address2.city);
-    }
-
-    if (address1.state && address2.state) {
-      addScore("State", address1.state, address2.state);
-    }
-
-    addScore("Country", address1.country, address2.country);
-    addScore("Zipcode", address1.zip, address2.zip);
-  }
-
-  // Calculate similarity for contact details
-  const contact1 = patient1.contact && patient1.contact.length > 0 ? patient1.contact[0] : null;
-  const contact2 = patient2.contact && patient2.contact.length > 0 ? patient2.contact[0] : null;
-
-  if (contact1 && contact2) {
-    if (contact1.phone && contact2.phone) {
-      addScore("Phone", contact1.phone, contact2.phone);
-    }
-
-    if (contact1.email && contact2.email) {
-      addScore("Email", contact1.email, contact2.email);
-    }
-  }
-
-  const totalScore = score / fieldCount;
-  similarityScores["Total Score"] = [totalScore];
-  console.log(JSON.stringify(similarityScores, null, 2)); // should this be logged using util.out with log or debug instead?
-  return totalScore;
 };
